Add relative path and put timeit decorator in helper module.
