--- simbev/results/log_file_simbev.txt
@@ -1,50 +1,0 @@
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0012281999999999016 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0012670000000001291 seconds 
-Function <function Trip._create_fast_charge_events at 0x000001A67FE49280> took 0.002511599999999836 seconds 
-Function <function Trip._create_fast_charge_events at 0x000001A67FE49280> took 0.001212599999999897 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0014552000000001009 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0014632999999997232 seconds 
-Function <function Trip._create_fast_charge_events at 0x000002C9E93A9280> took 0.0012742000000001141 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0014482000000000106 seconds 
-Function <function Trip._create_fast_charge_events at 0x000001A67FE49280> took 0.001441900000000107 seconds 
-Function <function Trip._create_fast_charge_events at 0x000001A67FE49280> took 0.0010684999999996947 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.001201099999999844 seconds 
-Function <function Trip._create_fast_charge_events at 0x000002C9E93A9280> took 0.0013244999999999507 seconds 
-Function <function Trip._create_fast_charge_events at 0x000001A67FE49280> took 0.001241100000000106 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0013068000000000524 seconds 
-Function <function Trip._create_fast_charge_events at 0x000001A67FE49280> took 0.001382700000000181 seconds 
-Function <function Trip._create_fast_charge_events at 0x000002C9E93A9280> took 0.001393100000000036 seconds 
-Function <function Trip._create_fast_charge_events at 0x000002C9E93A9280> took 0.0012598000000005882 seconds 
-Function <function Trip._create_fast_charge_events at 0x000002C9E93A9280> took 0.0013201000000000462 seconds 
-Function <function Trip._create_fast_charge_events at 0x000002C9E93A9280> took 0.0015292000000002304 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0013113000000002373 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0013154000000001886 seconds 
-Function <function Trip._create_fast_charge_events at 0x000001A67FE49280> took 0.002775299999999703 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0015122000000005187 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.001167200000000257 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0012470999999996124 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0025017000000007172 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0012706999999991808 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0013079999999998648 seconds 
-Function <function Trip._create_fast_charge_events at 0x000002C9E93A9280> took 0.0012900000000009015 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0012483999999997053 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0012777999999986633 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0035799000000000802 seconds 
-Function <function Trip._create_fast_charge_events at 0x000002C9E93A9280> took 0.001215200000000749 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.001421399999999906 seconds 
-Function <function Trip._create_fast_charge_events at 0x000002C9E93A9280> took 0.0012467999999987711 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0010162000000004667 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0011764999999996917 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0010778000000009058 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0010250000000002757 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0011067000000011262 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.001340800000001252 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0009246999999987793 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0012284000000004625 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0012980999999996357 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0011943999999992627 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0010608000000011941 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0010086000000004702 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000017A04AF9280> took 0.0010220999999983604 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0009381000000008299 seconds 
-Function <function Trip._create_fast_charge_events at 0x0000023D22F29280> took 0.0009473999999993765 seconds
